+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Features
 - Ability to create url custom fields. (thanks to [@astagi](https://github.com/astagi)).
+- Blocked projects support
 - Moved from iconfont to SVG sprite icon system and redesign.
 - Redesign 'Admin > Project > Modules' panel.
 - Add badge to project owners